# swagger-codegen/go-server/go/model_user.go
/*
 * Store API
 *
 * Store API documentation
 *
 * API version: 0.0.1
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type User struct {

	Id int64 `json:"id,omitempty"`

	Login string `json:"login"`

	FirstName string `json:"firstName,omitempty"`

	LastName string `json:"lastName,omitempty"`

	Email string `json:"email,omitempty"`

	Activated bool `json:"activated"`

	LangKey string `json:"langKey,omitempty"`

	ImageUrl string `json:"imageUrl,omitempty"`

	ResetDate time.Time `json:"resetDate,omitempty"`
}